fixed a lil bug
so a small bug i encountered was that whenever the values that were to be typewritten by the pyautogui library, it would end up joining the previous and current values together. i added a new line, which basically made it so that all the values were equal, and were reset everytime the function was called.